+++ project-egg-core/src/Core/Renderer.cpp
@@ -24,10 +24,7 @@ SDL_Renderer* Renderer::GetInstance() const {
 }
 
 void Renderer::Render() {
-    SDL_SetRenderDrawColor(s_Renderer, 124, 218, 254, 255);
-    SDL_RenderClear(s_Renderer);
-    AssetManager::GetInstance()->DrawTexture(s_Renderer, "ghost", 200, 50, 600, 660);
-    SDL_RenderPresent(s_Renderer);
+    
 }
 
 void Renderer::fillRect(SDL_Rect* rect)